re-organized bytes UTs
