+++ mod/primitives/pkg/bytes/b32_test.go
@@ -18,6 +18,7 @@
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
 // TITLE.
 
+//nolint:lll // long strings.
 package bytes_test
 
 import (
@@ -26,6 +27,245 @@ import (
 	"github.com/berachain/beacon-kit/mod/primitives/pkg/bytes"
 	"github.com/stretchr/testify/require"
 )
+
+func TestBytes32UnmarshalText(t *testing.T) {
+	tests := []struct {
+		name    string
+		input   string
+		want    bytes.B32
+		wantErr bool
+	}{
+		{
+			name:  "valid input",
+			input: "0x0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f20",
+			want: bytes.B32{
+				0x01,
+				0x02,
+				0x03,
+				0x04,
+				0x05,
+				0x06,
+				0x07,
+				0x08,
+				0x09,
+				0x0a,
+				0x0b,
+				0x0c,
+				0x0d,
+				0x0e,
+				0x0f,
+				0x10,
+				0x11,
+				0x12,
+				0x13,
+				0x14,
+				0x15,
+				0x16,
+				0x17,
+				0x18,
+				0x19,
+				0x1a,
+				0x1b,
+				0x1c,
+				0x1d,
+				0x1e,
+				0x1f,
+				0x20,
+			},
+			wantErr: false,
+		},
+		{
+			name:    "invalid input - not hex",
+			input:   "0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f20",
+			wantErr: true,
+		},
+		{
+			name:    "invalid input - wrong length",
+			input:   "0x0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f",
+			wantErr: true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			var got bytes.B32
+			err := got.UnmarshalText([]byte(tt.input))
+			if tt.wantErr {
+				require.Error(t, err, "Test case %s", tt.name)
+			} else {
+				require.NoError(t, err)
+				require.Equal(t, tt.want, got, "Test case %s", tt.name)
+			}
+		})
+	}
+}
+
+func TestBytes32UnmarshalJSON(t *testing.T) {
+	tests := []struct {
+		name    string
+		input   string
+		want    bytes.B32
+		wantErr bool
+	}{
+		{
+			name:  "valid input",
+			input: `"0x0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f20"`,
+			want: bytes.B32{
+				0x01,
+				0x02,
+				0x03,
+				0x04,
+				0x05,
+				0x06,
+				0x07,
+				0x08,
+				0x09,
+				0x0a,
+				0x0b,
+				0x0c,
+				0x0d,
+				0x0e,
+				0x0f,
+				0x10,
+				0x11,
+				0x12,
+				0x13,
+				0x14,
+				0x15,
+				0x16,
+				0x17,
+				0x18,
+				0x19,
+				0x1a,
+				0x1b,
+				0x1c,
+				0x1d,
+				0x1e,
+				0x1f,
+				0x20,
+			},
+			wantErr: false,
+		},
+		{
+			name:    "invalid input - not hex",
+			input:   `"0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f20"`,
+			wantErr: true,
+		},
+		{
+			name:    "invalid input - wrong length",
+			input:   `"0x0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f"`,
+			wantErr: true,
+		},
+		{
+			name:    "invalid input - extra characters",
+			input:   `"0x0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f202122"`,
+			wantErr: true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			var got bytes.B32
+			err := got.UnmarshalJSON([]byte(tt.input))
+			if tt.wantErr {
+				require.Error(t, err, "Test case: %s", tt.name)
+			} else {
+				require.NoError(t, err)
+				require.Equal(t, tt.want, got, "Test case: %s", tt.name)
+			}
+		})
+	}
+}
+
+func TestBytes32MarshalText(t *testing.T) {
+	tests := []struct {
+		name  string
+		input bytes.B32
+		want  string
+	}{
+		{
+			name: "valid input",
+			input: bytes.B32{0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
+				0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10, 0x11, 0x12,
+				0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d,
+				0x1e, 0x1f, 0x20},
+			want: "0x0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f20",
+		},
+		{
+			name:  "empty input",
+			input: bytes.B32{},
+			want:  "0x0000000000000000000000000000000000000000000000000000000000000000",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := tt.input.MarshalText()
+			require.NoError(t, err, "Test case: %s", tt.name)
+			require.Equal(t, tt.want, string(got), "Test case: %s", tt.name)
+		})
+	}
+}
+
+func TestBytes32String(t *testing.T) {
+	tests := []struct {
+		name  string
+		input bytes.B32
+		want  string
+	}{
+		{
+			name: "valid input",
+			input: bytes.B32{0x01, 0x02, 0x03, 0x04, 0x05, 0x06,
+				0x07, 0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10,
+				0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1a,
+				0x1b, 0x1c, 0x1d, 0x1e, 0x1f, 0x20},
+			want: "0x0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f20",
+		},
+		{
+			name:  "empty input",
+			input: bytes.B32{},
+			want:  "0x0000000000000000000000000000000000000000000000000000000000000000",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := tt.input.String()
+			require.Equal(t, tt.want, got, "Test case: %s", tt.name)
+		})
+	}
+}
+
+func TestToBytes32(t *testing.T) {
+	tests := []struct {
+		name     string
+		input    []byte
+		expected bytes.B32
+	}{
+		{
+			name:     "Input less than 32 bytes",
+			input:    []byte{1, 2, 3},
+			expected: bytes.B32{1, 2, 3},
+		},
+		{
+			name:     "Input exactly 32 bytes",
+			input:    make([]byte, 32),
+			expected: bytes.B32{},
+		},
+		{
+			name:     "Input more than 32 bytes",
+			input:    make([]byte, 40),
+			expected: bytes.B32{},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result := bytes.ToBytes32(tt.input)
+			require.Equal(t, tt.expected, result, "Test case: %s", tt.name)
+		})
+	}
+}
 
 func TestB32MarshalSSZ(t *testing.T) {
 	tests := []struct {
